# CRM/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CombatDen API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/members/": {
      "post": {
        "tags": [
          "members"
        ],
        "summary": "Create a member (provisions a Stripe customer)",
        "description": "Create a member and provision its Stripe customer. Gym staff only.\n\nEvery member is created with a Stripe customer. The gym must have a\nStripe Connect account; otherwise the request is rejected (400) and no\nmember row is written.",
        "operationId": "create_member_api_v1_members__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Member created with a Stripe customer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembersBillingProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, or gym has no Stripe account"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/members/{member_id}": {
      "put": {
        "tags": [
          "members"
        ],
        "summary": "Update a member",
        "description": "Update a member's mutable fields.",
        "operationId": "update_member_api_v1_members__member_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Member Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Member updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid update"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized to update this member"
          },
          "404": {
            "description": "Member not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "members"
        ],
        "summary": "Get member detail",
        "description": "Full CRM member detail for the Specific Member screen: all memberships (grouped by plan), linked family accounts, payment history, redeemed rewards, and the card on file. Membership status is derived from member_memberships, so a member with no membership resolves cleanly without error.",
        "operationId": "get_member_detail_api_v1_members__member_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Member Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member detail retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberBillingDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized to view this member"
          },
          "404": {
            "description": "Member not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/membership_plans/": {
      "post": {
        "tags": [
          "membership_plans"
        ],
        "summary": "Create a membership plan",
        "description": "Creates a membership plan with an initial price in Stripe and CRM.",
        "operationId": "create_plan_api_v1_membership_plans__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipPlanCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plan created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipPlanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "502": {
            "description": "Stripe error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "membership_plans"
        ],
        "summary": "Update a membership plan",
        "description": "Updates plan metadata in Stripe and CRM.",
        "operationId": "update_plan_api_v1_membership_plans__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipPlanUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipPlanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "404": {
            "description": "Plan not found"
          },
          "502": {
            "description": "Stripe error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "membership_plans"
        ],
        "summary": "Delete a membership plan",
        "description": "Soft-deletes a membership plan and deactivates its Stripe product.",
        "operationId": "delete_plan_api_v1_membership_plans__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Plan Id"
            }
          },
          {
            "name": "gym_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gym Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Plan deleted successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "404": {
            "description": "Plan not found"
          },
          "502": {
            "description": "Stripe error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "membership_plans"
        ],
        "summary": "List membership plans",
        "description": "Lists all non-deleted plans for a gym with their active prices.",
        "operationId": "list_plans_api_v1_membership_plans__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "gym_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gym Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plans listed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MembershipPlanResponse"
                  },
                  "title": "Response List Plans Api V1 Membership Plans  Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/membership_plans/{plan_id}": {
      "get": {
        "tags": [
          "membership_plans"
        ],
        "summary": "Get a membership plan",
        "description": "Fetches a single plan with its active price.",
        "operationId": "get_plan_api_v1_membership_plans__plan_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Plan Id"
            }
          },
          {
            "name": "gym_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gym Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipPlanResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "404": {
            "description": "Plan not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/membership_plans/price": {
      "post": {
        "tags": [
          "membership_plans"
        ],
        "summary": "Set plan price",
        "description": "Creates a new price for a plan. Existing members keep their old price \u2014 use the migrate endpoints to move them.",
        "operationId": "set_price_api_v1_membership_plans_price_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipPlanPriceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Price set successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipPlanPriceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "404": {
            "description": "Plan not found"
          },
          "502": {
            "description": "Stripe error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemberUpdateRequest": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MemberUpdateData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "MemberUpdateRequest",
        "description": "Body for PUT /api/v1/members/{member_id}."
      },
...
      "MembershipPlanPriceRequest": {
        "properties": {
          "plan_id": {
            "type": "string",
            "format": "uuid",
            "title": "Plan Id"
          },
          "gym_id": {
            "type": "string",
            "format": "uuid",
            "title": "Gym Id"
          },
          "price": {
            "type": "integer",
            "title": "Price"
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "gym_id",
          "price"
        ],
        "title": "MembershipPlanPriceRequest",
        "description": "Set / update the active price on a plan."
      },
...
      "MembershipPlanUpdateRequest": {
        "properties": {
          "plan_id": {
            "type": "string",
            "format": "uuid",
            "title": "Plan Id"
          },
          "gym_id": {
            "type": "string",
            "format": "uuid",
            "title": "Gym Id"
          },
          "data": {
            "$ref": "#/components/schemas/MembershipPlanUpdateData"
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "gym_id",
          "data"
        ],
        "title": "MembershipPlanUpdateRequest",
        "description": "Update a membership plan. Separates IDs from mutable data."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}